requirements.txt
openai==0.28.0
scrapy==2.11.2
pymongo==4.10.1
python-telegram-bot==13.7
python-dotenv==1.0.0